refactor(app): rename components
