+++ packages/walletts-app/app/renderer/store/account/state/index.ts
@@ -15,8 +15,8 @@ export const defaultAccounts: IAccountState = {
       id: "defaultaccountid",
       iconUrl: "fas fa-wallet",
       webview: null,
-      member: defaultUserIDs,
-      owners: defaultKnownUserIDs,
+      member: defaultUserIDs.slice(0, 1),
+      owners: defaultUserIDs.slice(0, 1),
       balance: Satoshi.fromNumber(1000).value,
       isActive: true,
       integratedChannels: defaultChannelIDs,
